+++ src/type.ts
@@ -1,0 +1,8 @@
+type ThumbnailCache = Record<string, Thumbnail>;
+
+type Thumbnail = {
+  path: string;
+  updated_time: number;
+};
+
+export { ThumbnailCache, Thumbnail };